--- src/components/MyInput/index.jsx
@@ -1,3 +1,0 @@
-export default function MyInput() {
-  return <div>인풋예시</div>;
-}
