+++ src/pages/RegionPage/RegionPage.jsx
@@ -2,10 +2,11 @@ import Divider from '../../components/Divider';
 import MainContainer from '../../components/MainContainer';
 import RankTitle from '@components/RankTitle';
 import MyProfile from '../../components/MyProfile';
-import RegionRankList from './RegionRankList';
+
 import { MyRegionCard, RegionWinnerCard } from '@components/Cards';
 
 import level4 from './../../assets/images/regionLevel/regionLevel4.svg'; //임시
+import RegionRankList from './RegionRankList/RegionRankList';
 
 export default function RegionPage() {
 	return (
