+++ src/routes/mainRoutes.jsx
@@ -4,11 +4,11 @@ import MainLayout from '../layout/MainLayout';
 import { routes } from '../constants/routes';
 import MainPage from '../pages/MainPage';
 import SchoolPage from '../pages/SchoolPage';
-import RegionPage from '../pages/RegionPage';
 import MyPage from '../pages/MyPage';
 import AboutPage from '../pages/AboutPage';
 import PrivacyPage from '../pages/TermsPage/PrivacyPage';
 import TosPage from '../pages/TermsPage/TosPage';
+import RegionPage from '../pages/RegionPage/RegionPage';
 
 export const mainRoutes = {
 	path: '/',
